+++ tokio/src/net/tcp/listener.rs
@@ -1,6 +1,6 @@
 use crate::future::poll_fn;
 use crate::net::tcp::{Incoming, TcpStream};
-use crate::net::util::IoSource;
+use crate::net::util::IoResource;
 use crate::net::ToSocketAddrs;
 
 use std::convert::TryFrom;
@@ -30,7 +30,7 @@ use std::task::{Context, Poll};
 /// }
 /// ```
 pub struct TcpListener {
-    io: IoSource<mio::net::TcpListener>,
+    io: IoResource<mio::net::TcpListener>,
 }
 
 impl TcpListener {
@@ -182,12 +182,12 @@ impl TcpListener {
     /// ```
     pub fn from_std(listener: net::TcpListener) -> io::Result<TcpListener> {
         let io = mio::net::TcpListener::from_std(listener);
-        let io = IoSource::new(io)?;
+        let io = IoResource::new(io)?;
         Ok(TcpListener { io })
     }
 
     fn new(listener: mio::net::TcpListener) -> io::Result<TcpListener> {
-        let io = IoSource::new(listener)?;
+        let io = IoResource::new(listener)?;
         Ok(TcpListener { io })
     }
