+++ tokio/src/net/util/mod.rs
@@ -1,4 +1,4 @@
 //! Utilities for implementing networking types.
 
-mod io_source;
-pub use self::io_source::IoSource;
+mod io_resource;
+pub use self::io_resource::IoResource;
